Moved LoginAttempts model class out of model.py
